requirements.txt
discord.py==2.4.0
Flask==3.0.0
